+++ demo/discrete-cubism.html
@@ -15,8 +15,9 @@
 <button id="AddEvent" onClick="javascript:addEvent();">Go!</button>
 <br/>
 <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
-<script src="http://square.github.io/cubism/d3.v2.js"></script>
-<script src="http://square.github.io/cubism/cubism.v1.js"></script>
+<script src="./lib/d3.v5.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/es6-shim/0.35.3/es6-shim.js"></script>
+<script src="./cubism-es.standalone.js"></script>
 <script src="http://square.github.io/cubism/highlight.min.js"></script>
 <script src="discrete-cubism.js"></script>
 </BODY>
